Fixed logging.

+++ Data/Matlab/MatlabValue.cs
@@ -6,7 +6,7 @@
         public string Symbol { get; set; }
         public double Value { get; set; }
         public string Account { get; set; }
-        public double Equity { get; set; }
+        public decimal Equity { get; set; }
         public MatlabType ValueType { get; set; }
         public string Name { get; set; }
     }
